requirements.txt
azure-functions
adal
msrestazure
azure.mgmt.media